refactor(backend): replace mock EmailService with real MailService using JavaMailSender
