+++ backend/src/main/java/com/cinerama/backend/service/MailService.java
@@ -1,5 +1,5 @@
 package com.cinerama.backend.service;
 
 public interface MailService {
-    void sendEmail(String to, String subject, String body);
+    void sendVerificationEmail(String toEmail, String verificationCode);
 }
