+++ backend/src/main/java/com/cinerama/backend/service/impl/AuthServiceImpl.java
@@ -6,7 +6,7 @@ import com.cinerama.backend.dto.VerificationRequest;
 import com.cinerama.backend.entity.User;
 import com.cinerama.backend.repository.UserRepository;
 import com.cinerama.backend.service.AuthService;
-import com.cinerama.backend.util.EmailService;
+import com.cinerama.backend.service.MailService;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -21,38 +21,45 @@ public class AuthServiceImpl implements AuthService {
 
     private final UserRepository userRepository;
     private final BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
-    private final EmailService emailService;
+    private final MailService mailService;
 
     @Override
     public User register(RegisterRequest request) {
+        // Check if the provided passwords match
         if (!request.getPassword().equals(request.getConfirmPassword())) {
             throw new IllegalArgumentException("Passwords do not match");
         }
 
+        // Create a new user entity with the provided details
         User user = User.builder()
                 .name(request.getName())
                 .email(request.getEmail())
-                .password(passwordEncoder.encode(request.getPassword()))
-                .enabled(false)
-                .verificationCode(UUID.randomUUID().toString())
+                .password(passwordEncoder.encode(request.getPassword())) // Encrypt the password
+                .enabled(false) // Set the account as disabled until verification
+                .verificationCode(UUID.randomUUID().toString()) // Generate a unique verification code
                 .build();
 
+        // Save the user to the database
         User savedUser = userRepository.save(user);
 
-        emailService.sendVerificationEmail(savedUser.getEmail(), savedUser.getVerificationCode());
+        // Send a verification email to the user
+        mailService.sendVerificationEmail(savedUser.getEmail(), savedUser.getVerificationCode());
 
         return savedUser;
     }
 
     @Override
     public void verify(VerificationRequest request) {
+        // Find the user by email or throw an exception if not found
         User user = userRepository.findByEmail(request.getEmail())
                 .orElseThrow(() -> new IllegalArgumentException("User not found"));
 
+        // Check if the provided verification code matches the one stored
         if (!request.getVerificationCode().equals(user.getVerificationCode())) {
             throw new IllegalArgumentException("Invalid verification code");
         }
 
+        // Enable the user account and clear the verification code
         user.setEnabled(true);
         user.setVerificationCode(null);
         userRepository.save(user);
@@ -60,13 +67,16 @@ public class AuthServiceImpl implements AuthService {
 
     @Override
     public User login(LoginRequest request) {
+        // Find the user by email or throw an exception if not found
         User user = userRepository.findByEmail(request.getEmail())
                 .orElseThrow(() -> new IllegalArgumentException("Invalid email or password"));
 
+        // Check if the provided password matches the stored encrypted password
         if (!passwordEncoder.matches(request.getPassword(), user.getPassword())) {
             throw new IllegalArgumentException("Invalid email or password");
         }
 
+        // Check if the user account is verified
         if (!user.isEnabled()) {
             throw new IllegalStateException("Account is not verified");
         }
